feat!: Network Time (RFC #14) (#845)
* refactor: simplify tick system, introduce server based time

* feat!: remove receive tick

* feat!: remove max received events per tick rate

* feat!: Update components to use new time system. Remove time sync message.

* test: NetworkTime operators and creation tests

* test: Upgrade manual network variable test to new time system

* docs: network config tooltip for tickrate

* feat!: completely remove old network time

* feat: decouple network time system

* feat: NetworkTime use double everywhere

* docs: Improve tick rate tool tip.

* change tickrate to 30

+++ com.unity.multiplayer.mlapi/Runtime/NetworkVariable/NetworkVariableSettings.cs
@@ -34,7 +34,7 @@ namespace MLAPI.NetworkVariable
         /// A value of 0 will cause the variable to sync as soon as possible after being changed.
         /// A value of less than 0 will cause the variable to sync only at once at spawn and not update again.
         /// </summary>
-        public float SendTickrate = 0;
+        public double SendTickrate = 0;
 
         /// <summary>
         /// The name of the channel to use for this variable.
